% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{run_IIDEA}
\alias{run_IIDEA}
\title{Open Rshiny app}
\usage{
run_IIDEA()
}
\value{
An object that represents the app. Printing the object or passing
it to runApp() will run the app.
}
\description{
Open Rshiny app
}
